Post-run cleanup batch: dtype rename, honest heartbeat, quiet pull, no dead checkpoints
All queued cleanup now that the run is complete (generation-path freeze lifted):
- _model.py: torch_dtype -> dtype (transformers deprecation warning gone).
- score_offline.py: heartbeat rate/ETA computed from GENERATED prompts only —
  resume-skipped alphas previously counted as done, making the first heartbeats
  absurd (117 gen/s decaying toward truth).
- gpu_job.sh + run_a100.sh factqa stage: drop --save_every from the calibration
  teach (its checkpoints were never used and 6x'd adapter storage — part of the
  43GB volume bloat). The HAZARD teach keeps --save_every 5: its checkpoint sweep
  is a reported result.
- colab_full_modal.ipynb: pull cell replaced with the quiet SDK pull (transcripts
  only, never the multi-GB adapters/, ~4 lines of output, zip straight to Drive).
- main.tex reproducibility: state the batching self-check precisely — at most one
  diverging probe completion tolerated, otherwise single-stream fallback (matches
  the deployed tolerance=1 exactly).

+++ experiments/unlearning/_model.py
@@ -26,6 +26,6 @@ def load_base(model_id: str, *, load_4bit: bool, device: str, dtype=torch.float3
         # `--device cuda:1` doesn't load onto GPU 0 and fail the first forward with a
         # cross-device error. ("cuda" -> current device; an explicit "cuda:N" is honored.)
         return AutoModelForCausalLM.from_pretrained(
-            model_id, quantization_config=bnb, device_map={"": device}, torch_dtype=torch.bfloat16,
+            model_id, quantization_config=bnb, device_map={"": device}, dtype=torch.bfloat16,
         )
-    return AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=dtype).to(device)
+    return AutoModelForCausalLM.from_pretrained(model_id, dtype=dtype).to(device)
